Tidy up FloorItem
