+++ engine/src/main/kotlin/world/gregs/voidps/engine/entity/item/floor/FloorItem.kt
@@ -10,27 +10,25 @@ import world.gregs.voidps.engine.map.Tile
  * An [Item] with physical location
  * Not a data class to prevent hash conflicts in lists
  */
-class FloorItem private constructor(
+class FloorItem(
     val id: String,
     override var tile: Tile,
-    var amount: Int,
-    override val size: Size,
-    var owner: Int
+    var amount: Int = 1,
+    var disappearTimer: Int = -1,
+    var revealTimer: Int = -1,
+    var owner: Int = 0
 ) : Entity {
 
-    constructor(id: String, tile: Tile, amount: Int = 1, disappear: Int = -1, reveal: Int = -1, owner: Int = 0) : this(id, tile, amount, Size.ONE, owner) {
-        this.disappearTimer = disappear
-        this.revealTimer = reveal
-    }
-
+    override val size: Size = Size.ONE
     override val events: Events = Events(this)
     lateinit var def: ItemDefinition
-    var disappearTimer: Int = -1
-    var revealTimer: Int = -1
 
     val value: Long
         get() = def.cost * amount.toLong()
 
+    /**
+     * Adds [other] items amount to this item.
+     */
     fun combine(other: FloorItem): Boolean {
         if (def.stackable != 1) {
             return false
